GCS Master Code/requirements.txt
pyqt5
pyqtgraph
pyserial>=3.5
PyQt5>=5.15.9
matplotlib>=3.7.1
numpy>=1.24.3
# requirements.txt

fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pyserial>=3.5
matplotlib>=3.7.1
requests>=2.31.0

